Fix inquiry details and delete inquiry view only owner can access it

+++ inquiries/views.py
@@ -44,9 +44,17 @@ def inquiry_details(request, inquiry_id):
     This view displays the inquiry details and
     allows user to add an update/reply
     to the same inquiry.
+    The view checks if the users is the 
+    owner of the inquiry
     """
 
     inquiry = get_object_or_404(Inquiry, pk=inquiry_id)
+
+    if request.user != inquiry.user:
+        messages.error(
+                    request, 
+                    "You don't have permission to access this inquiry.")
+        return redirect('home')
 
     # Split user_reply into a list of replies
     user_replies = inquiry.user_reply.split('\n') if inquiry.user_reply else []
@@ -77,8 +85,16 @@ def inquiry_details(request, inquiry_id):
 def delete_inquiry(request, inquiry_id):
     """
     This deletes the inquiry.
+    The view checks if the users is the 
+    owner of the inquiry
     """
     inquiry = get_object_or_404(Inquiry, pk=inquiry_id)
+
+    if request.user != inquiry.user:
+        messages.error(
+                    request, 
+                    "You don't have permission to access this inquiry.")
+        return redirect('home')
 
     if request.method == 'POST':
         inquiry.delete()
